refactor: Update blog page metadata generation
This commit updates the metadata generation for the blog page. It fixes an issue where the title and description were not being properly set when fetching the blog post data. Now, if the fetch is successful, the title and description are set to the values from the fetched data. Otherwise, the title is set to "preview blog post" and the description is left empty.

The code changes also include accessing and extending the parent metadata, as well as updating the current image URL.