refactor: Update SingleBlogPost component to conditionally render author information
This commit updates the SingleBlogPost component to conditionally render the author information. Previously, the author's image and name were always displayed, even if the author data was not available. Now, the author information is only rendered if the author object exists.

The code changes include wrapping the author information in a conditional statement and checking if the author object is truthy before rendering the image and name.

Refactor the SingleBlogPost component to improve rendering of author information